+++ src/compas/datastructures/tree/tree.py
@@ -41,35 +41,46 @@ class TreeNode(Data):
         "type": "object",
         "$recursiveAnchor": True,
         "properties": {
+            "name": {"type": "string"},
             "attributes": {"type": "object"},
             "children": {"type": "array", "items": {"$recursiveRef": "#"}},
         },
-        "required": ["attributes", "children"],
     }
 
     @property
     def __data__(self):
-        return {
-            "attributes": self.attributes,
-            "children": [child.__data__ for child in self.children],
-        }
+        data = {}
+        if self.name is not None:
+            data["name"] = self.name
+        if self.attributes:
+            data["attributes"] = self.attributes
+        if self.children:
+            data["children"] = [child.__data__ for child in self.children]
+        return data
 
     @classmethod
     def __from_data__(cls, data):
-        node = cls(**data["attributes"])
-        for child in data["children"]:
+
+        name = data.get("name", None)
+        attributes = data.get("attributes", {})
+        children = data.get("children", [])
+
+        node = cls(name=name, **attributes)
+        for child in children:
             node.add(cls.__from_data__(child))
         return node
 
-    def __init__(self, **kwargs):
-        super(TreeNode, self).__init__(**kwargs)
+    def __init__(self, name=None, **kwargs):
+        super(TreeNode, self).__init__(name=name, **kwargs)
         self.attributes = kwargs
         self._parent = None
         self._children = []
         self._tree = None
 
     def __repr__(self):
-        return "<TreeNode {}>".format(self.name)
+        if self._name:
+            return "<TreeNode: {}>".format(self._name)
+        return "<TreeNode>"
 
     @property
     def is_root(self):
@@ -427,14 +438,65 @@ class Tree(Datastructure):
     def __repr__(self):
         return "<Tree with {} nodes>".format(len(list(self.nodes)))
 
-    def print_hierarchy(self):
-        """Print the spatial hierarchy of the tree."""
+    def print_hierarchy(self, max_depth=None):
+        """
+        Print the spatial hierarchy of the tree.
 
-        def _print(node, prefix="", last=True):
+        Parameters
+        ----------
+        max_depth : int, optional
+            The maximum depth of the hierarchy to print.
+            Default is ``None``, in which case the entire hierarchy is printed.
+
+        Returns
+        -------
+        None
+
+        """
+
+        def _print(node, prefix="", last=True, depth=0):
+
+            if max_depth is not None and depth > max_depth:
+                return
+
             connector = "└── " if last else "├── "
             print("{}{}{}".format(prefix, connector, node))
             prefix += "    " if last else "│   "
             for i, child in enumerate(node.children):
-                _print(child, prefix, i == len(node.children) - 1)
+                _print(child, prefix, i == len(node.children) - 1, depth + 1)
 
         _print(self.root)
+
+    def to_graph(self, key_mapper=None):
+        """Convert the tree to a graph.
+
+        Parameters
+        ----------
+        key_mapper : callable, optional
+            A callable to map the tree node to a key in the graph.
+            Default is ``None``, in which case the index of the node is used.
+
+        Returns
+        -------
+        :class:`compas.datastructures.Graph`
+            The graph.
+
+        """
+        from compas.datastructures import Graph
+
+        graph = Graph(**self.attributes)
+        nodes = list(self.nodes)
+
+        if key_mapper is None:
+            key_mapper = lambda node: nodes.index(node)  # noqa: E731
+
+        for node in nodes:
+            graph.add_node(key=key_mapper(node), attr_dict=node.attributes, name=node._name)
+
+        for node in self.nodes:
+            if node.parent:
+                u = key_mapper(node.parent)
+                v = key_mapper(node)
+                graph.add_edge(u, v)
+
+        return graph